refactor: 각 페이지 별 컴포넌트 파일 및 상태 분리
