+++ src/store/store.js
@@ -1,0 +1,10 @@
+const store = {
+  isLoggedIn: false,
+  user: {
+    username: "",
+    email: "",
+    bio: "",
+  },
+};
+
+export default store;
